Move QueryResult interface to a separate file
